feat: add cdr command
the cdr command will change to <index> directory from a list of 10
most recently moved to directories. this only works for the interactive
cd command, and not the fs.cd function.
you can find the list of recent directories with `cdr list`.
usage: `cdr <index>`
the `cdr help` command also gives this bit of info