feat!: enhance docs and doc command
changes the actual file format of docs
to markup since that's basically what we have been
using in the first place.

the docgen command has been modified to write
markdown headings with the function name and
yaml metadata for easy consumption by hugo for the
website (soon).

all other docs have been moved to markdown as well
this is the main reason this is a "breaking" change
users will have to reinstall hilbish
(task uninstall and task install) to remove the old
plaintext docs